flatten scraper modules
