--- scraper/barbora/src/main/java/GroceryFamily/GroceryDad/barbora/productCheckOnThePage/PriceChecking.java
@@ -1,4 +1,0 @@
-package GroceryFamily.GroceryDad.barbora.productCheckOnThePage;
-
-public class PriceChecking {
-}
