--- scraper/barbora/src/test/java/GroceryFamily/GroceryDad/barbora/BarboraScraperTest.java
@@ -1,14 +1,0 @@
-package GroceryFamily.GroceryDad.barbora;
-
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@Disabled
-@SpringBootTest(classes = BarboraScraper.class)
-class BarboraScraperTest {
-    @Test
-    void test() {
-        // todo: implement some better test
-    }
-}
