--- scraper/core/pom.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>GroceryFamily.GroceryDad</groupId>
-        <artifactId>GroceryDad</artifactId>
-        <version>0.1.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>core</artifactId>
-
-    <dependencies>
-        <!-- https://github.com/GroceryFamily -->
-        <dependency>
-            <groupId>GroceryFamily</groupId>
-            <artifactId>GroceryElders</artifactId>
-        </dependency>
-
-        <!-- https://github.com/selenide/selenide -->
-        <dependency>
-            <groupId>com.codeborne</groupId>
-            <artifactId>selenide</artifactId>
-        </dependency>
-
-        <!-- https://github.com/antivoland/simple-file-cache -->
-        <dependency>
-            <groupId>io.github.antivoland</groupId>
-            <artifactId>simple-file-cache</artifactId>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <resources>
-            <resource>
-                <directory>src/main/resources</directory>
-                <filtering>true</filtering>
-            </resource>
-        </resources>
-
-        <plugins>
-            <plugin>
-                <groupId>org.springframework.boot</groupId>
-                <artifactId>spring-boot-maven-plugin</artifactId>
-            </plugin>
-        </plugins>
-    </build>
-</project>
