--- scraper/prisma/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-spring.profiles.active=@spring.profiles.active@
-
-scraper.prisma.cache.directory=cache/prisma
