--- scraper/rimi/src/main/java/GroceryFamily/GroceryDad/rimi/Config.java
@@ -1,19 +1,0 @@
-package GroceryFamily.GroceryDad.rimi;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-
-@Configuration
-@Import(value = {GroceryFamily.GroceryElders.Config.class})
-public class Config {
-    @Bean
-    WebDriver webDriver() {
-        ChromeOptions options = new ChromeOptions();
-        options.addArguments("--start-maximized");
-        return new ChromeDriver(options);
-    }
-}
