--- scraper/rimi/src/main/resources/application.yaml
@@ -1,8 +1,0 @@
-spring.profiles.active: @spring.profiles.active@
-
-GroceryDad:
-  scrapers:
-    rimi:
-      uri: https://rimi.ee/epood/en
-      cache:
-        directory: cache/rimi
