+++ scraper/src/main/java/GroceryFamily/scraper/productCheckOnThePage/PriceChecking.java
@@ -1,0 +1,4 @@
+package GroceryFamily.scraper.productCheckOnThePage;
+
+public class PriceChecking {
+}
