+++ scraper/src/main/resources/application.yaml
@@ -1,0 +1,16 @@
+spring.profiles.active: @spring.profiles.active@
+
+GroceryDad:
+  scrapers:
+    barbora:
+      uri: https://barbora.ee
+      cache:
+        directory: cache/barbora
+    prisma:
+      uri: https://prismamarket.ee
+      cache:
+        directory: cache/prisma
+    rimi:
+      uri: https://rimi.ee/epood/en
+      cache:
+        directory: cache/rimi
