+++ scraper/src/test/java/GroceryFamily/GroceryDad/ScraperTest.java
@@ -1,0 +1,15 @@
+package GroceryFamily.GroceryDad;
+
+import GroceryFamily.scraper.BarboraScraper;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@Disabled
+@SpringBootTest(classes = BarboraScraper.class)
+class ScraperTest {
+    @Test
+    void test() {
+        // todo: implement some better test
+    }
+}
